integrate schema, column, table into schema.go
